+++ packages/appcd-core/src/server.js
@@ -25,13 +25,24 @@ import { purgeUnusedNodejsExecutables } from 'appcd-nodejs';
 const { __n } = i18n();
 
 const logger = appcdCoreLogger('appcd:server');
-const { highlight, notice } = appcdCoreLogger.styles;
+const { highlight, note, notice } = appcdCoreLogger.styles;
 
 /**
  * The main server logic for the Appc Daemon. It controls all core aspects of the daemon including
  * plugins, logging, and request dispatching.
  */
 export default class Server {
+	static STOPPED  = 'stopped';
+	static STARTING = 'starting';
+	static STARTED  = 'started';
+	static STOPPING = 'stopping';
+
+	/**
+	 * The server state.
+	 * @type {String}
+	 */
+	state = Server.STOPPED;
+
 	/**
 	 * Creates a server instance and loads the configuration.
 	 *
@@ -76,6 +87,8 @@ export default class Server {
 	 * @access public
 	 */
 	async start() {
+		this.state = Server.STARTING;
+
 		const uid = this.config.get('server.user');
 		const gid = this.config.get('server.group');
 
@@ -141,13 +154,19 @@ export default class Server {
 			});
 
 		// listen for CTRL-C and SIGTERM
-		const shutdown = async () => {
-			try {
-				await this.shutdown();
-				process.exit(0);
-			} catch (err) {
-				logger.error(err);
+		const shutdown = async signal => {
+			if (this.state === Server.STARTED || this.state === Server.STARTING) {
+				logger.log(`Received signal ${highlight(signal)}, shutting down ${note(`(state=${this.state})`)}`);
+				try {
+					await this.shutdown();
+					process.exit(0);
+				} catch (err) {
+					logger.error(err);
+				}
+			} else if (this.state === Server.STOPPED) {
+				logger.log(`Received signal ${highlight(signal)}, but server already stopped ${note(`(state=${this.state})`)}`);
 			}
+			// if state === stopping, then we just ignore it
 		};
 		process.on('SIGINT',  shutdown);
 		process.on('SIGTERM', shutdown);
@@ -222,7 +241,7 @@ export default class Server {
 		// init the plugin manager
 		logger.log(`Initializing plugin system (api version ${appcdPluginAPIVersion})`);
 		this.systems.pluginManager = await new PluginManager({
-			paths: getPluginPaths(this.config.get('home'))
+			paths: getPluginPaths(homeDir)
 		}).init();
 
 		Dispatcher.register('/appcd/plugin', this.systems.pluginManager);
@@ -303,6 +322,8 @@ export default class Server {
 		// cleanup unused Node.js executables every hour
 		this.unsuedNodeCleanupTimer = setInterval(() => this.purgeUnusedNodejs(), 60 * 60 * 1000);
 		this.purgeUnusedNodejs();
+
+		this.state = Server.STARTED;
 	}
 
 	/**
@@ -331,6 +352,7 @@ export default class Server {
 	 * @access public
 	 */
 	async shutdown() {
+		this.state = Server.STOPPING;
 		logger.log('Shutting down server gracefully');
 
 		if (this.unsuedNodeCleanupTimer) {
@@ -368,6 +390,7 @@ export default class Server {
 			}
 		}
 
+		this.state = Server.STOPPED;
 		logger.log('appcd shutdown successfully');
 		return this;
 	}