+++ packages/appcd-subprocess/CHANGELOG.md
@@ -1,6 +1,7 @@
 # v4.0.1
 
  * fix: Properly handle IPC send errors when child process goes away.
+ * fix: Kill non-detached child processes when response stream is closed by the caller.
 
 # v4.0.0 (Jun 12, 2020)